using castToGeneric and changing quick data

+++ lib/views/pc/widgets/navigation_tab.dart
@@ -58,7 +58,8 @@ class _NavigationTab extends State<NavigationTab> {
               Navigator.pushReplacement(
                 context,
                 MaterialPageRoute<CompareScreen<int>>(
-                  builder: (final BuildContext context) => CompareScreen<int>(),
+                  builder: (final BuildContext context) =>
+                      CompareScreen<double>(),
                 ),
               );
             },
